qwert

+++ getSomeImages.py
@@ -52,7 +52,7 @@ for j in [0, 1, 2]:
 	i = 0
 	while True:
 		try:
-			im_slice = TrainNo[yes_inds[j], i, :, :, :]
+			im_slice = TrainNo[yes_inds[j], i, :, :, :].astype("uint8")
 			print("no: ", im_slice.dtype)
 			im = Image.fromarray(im_slice)
 			im.save(os.path.join(SaveDir, "no" + str(j), str(i) + ".png"))
